+++ instrumentation/opentelemetry-instrumentation-dbapi/src/opentelemetry/instrumentation/dbapi/__init__.py
@@ -46,6 +46,7 @@ import wrapt
 from opentelemetry import trace as trace_api
 from opentelemetry.instrumentation.dbapi.version import __version__
 from opentelemetry.instrumentation.utils import unwrap
+from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.trace import SpanKind, TracerProvider, get_tracer
 from opentelemetry.trace.status import Status, StatusCode
 
@@ -227,20 +228,17 @@ class DatabaseApiIntegration:
             }
         self._name = name
         self._version = version
-        self._tracer_provider = tracer_provider
+        self._tracer = get_tracer(
+            self._name,
+            instrumenting_library_version=self._version,
+            tracer_provider=tracer_provider,
+        )
         self.capture_parameters = capture_parameters
         self.database_system = database_system
         self.connection_props = {}
         self.span_attributes = {}
         self.name = ""
         self.database = ""
-
-    def get_tracer(self):
-        return get_tracer(
-            self._name,
-            instrumenting_library_version=self._version,
-            tracer_provider=self._tracer_provider,
-        )
 
     def wrapped_connection(
         self,
@@ -278,13 +276,13 @@ class DatabaseApiIntegration:
         if user and isinstance(user, bytes):
             user = user.decode()
         if user is not None:
-            self.span_attributes["db.user"] = str(user)
+            self.span_attributes[SpanAttributes.DB_USER] = str(user)
         host = self.connection_props.get("host")
         if host is not None:
-            self.span_attributes["net.peer.name"] = host
+            self.span_attributes[SpanAttributes.NET_PEER_NAME] = host
         port = self.connection_props.get("port")
         if port is not None:
-            self.span_attributes["net.peer.port"] = port
+            self.span_attributes[SpanAttributes.NET_PEER_PORT] = port
 
 
 def get_traced_connection_proxy(
@@ -325,10 +323,12 @@ class CursorTracer:
             return
         statement = self.get_statement(cursor, args)
         span.set_attribute(
-            "db.system", self._db_api_integration.database_system
+            SpanAttributes.DB_SYSTEM, self._db_api_integration.database_system
         )
-        span.set_attribute("db.name", self._db_api_integration.database)
-        span.set_attribute("db.statement", statement)
+        span.set_attribute(
+            SpanAttributes.DB_NAME, self._db_api_integration.database
+        )
+        span.set_attribute(SpanAttributes.DB_STATEMENT, statement)
 
         for (
             attribute_key,
@@ -367,7 +367,7 @@ class CursorTracer:
                 else self._db_api_integration.name
             )
 
-        with self._db_api_integration.get_tracer().start_as_current_span(
+        with self._db_api_integration._tracer.start_as_current_span(
             name, kind=SpanKind.CLIENT
         ) as span:
             self._populate_span(span, cursor, *args)