+++ instrumentation/opentelemetry-instrumentation-pyramid/src/opentelemetry/instrumentation/pyramid/callbacks.py
@@ -21,8 +21,12 @@ from pyramid.tweens import EXCVIEW
 
 import opentelemetry.instrumentation.wsgi as otel_wsgi
 from opentelemetry import context, trace
+from opentelemetry.instrumentation.propagators import (
+    get_global_response_propagator,
+)
 from opentelemetry.instrumentation.pyramid.version import __version__
 from opentelemetry.propagate import extract
+from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.util._time import _time_ns
 from opentelemetry.util.http import get_excluded_urls
 
@@ -95,7 +99,9 @@ def _before_traversal(event):
     if span.is_recording():
         attributes = otel_wsgi.collect_request_attributes(request_environ)
         if request.matched_route:
-            attributes["http.route"] = request.matched_route.pattern
+            attributes[
+                SpanAttributes.HTTP_ROUTE
+            ] = request.matched_route.pattern
         for key, value in attributes.items():
             span.set_attribute(key, value)
 
@@ -156,6 +162,10 @@ def trace_tween_factory(handler, registry):
                     response_or_exception.status,
                     response_or_exception.headers,
                 )
+
+                propagator = get_global_response_propagator()
+                if propagator:
+                    propagator.inject(response.headers)
 
                 activation = request.environ.get(_ENVIRON_ACTIVATION_KEY)