+++ instrumentation/opentelemetry-instrumentation-tornado/tests/test_instrumentation.py
@@ -18,11 +18,17 @@ from unittest.mock import Mock, patch
 from tornado.testing import AsyncHTTPTestCase
 
 from opentelemetry import trace
+from opentelemetry.instrumentation.propagators import (
+    TraceResponsePropagator,
+    get_global_response_propagator,
+    set_global_response_propagator,
+)
 from opentelemetry.instrumentation.tornado import (
     TornadoInstrumentor,
     patch_handler_class,
     unpatch_handler_class,
 )
+from opentelemetry.semconv.trace import SpanAttributes
 from opentelemetry.test.test_base import TestBase
 from opentelemetry.trace import SpanKind
 from opentelemetry.util.http import get_excluded_urls, get_traced_request_attrs
@@ -42,7 +48,11 @@ class TornadoTest(AsyncHTTPTestCase, TestBase):
         return app
 
     def setUp(self):
-        TornadoInstrumentor().instrument()
+        TornadoInstrumentor().instrument(
+            server_request_hook=getattr(self, "server_request_hook", None),
+            client_request_hook=getattr(self, "client_request_hook", None),
+            client_response_hook=getattr(self, "client_response_hook", None),
+        )
         super().setUp()
         # pylint: disable=protected-access
         self.env_patch = patch.dict(
@@ -127,12 +137,13 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             server,
             {
-                "http.method": method,
-                "http.scheme": "http",
-                "http.host": "127.0.0.1:" + str(self.get_http_port()),
-                "http.target": "/",
-                "net.peer.ip": "127.0.0.1",
-                "http.status_code": 201,
+                SpanAttributes.HTTP_METHOD: method,
+                SpanAttributes.HTTP_SCHEME: "http",
+                SpanAttributes.HTTP_HOST: "127.0.0.1:"
+                + str(self.get_http_port()),
+                SpanAttributes.HTTP_TARGET: "/",
+                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_STATUS_CODE: 201,
             },
         )
 
@@ -143,9 +154,9 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             client,
             {
-                "http.url": self.get_url("/"),
-                "http.method": method,
-                "http.status_code": 201,
+                SpanAttributes.HTTP_URL: self.get_url("/"),
+                SpanAttributes.HTTP_METHOD: method,
+                SpanAttributes.HTTP_STATUS_CODE: 201,
             },
         )
 
@@ -199,12 +210,13 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             server,
             {
-                "http.method": "GET",
-                "http.scheme": "http",
-                "http.host": "127.0.0.1:" + str(self.get_http_port()),
-                "http.target": url,
-                "net.peer.ip": "127.0.0.1",
-                "http.status_code": 201,
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_SCHEME: "http",
+                SpanAttributes.HTTP_HOST: "127.0.0.1:"
+                + str(self.get_http_port()),
+                SpanAttributes.HTTP_TARGET: url,
+                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_STATUS_CODE: 201,
             },
         )
 
@@ -215,9 +227,9 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             client,
             {
-                "http.url": self.get_url(url),
-                "http.method": "GET",
-                "http.status_code": 201,
+                SpanAttributes.HTTP_URL: self.get_url(url),
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_STATUS_CODE: 201,
             },
         )
 
@@ -234,12 +246,13 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             server,
             {
-                "http.method": "GET",
-                "http.scheme": "http",
-                "http.host": "127.0.0.1:" + str(self.get_http_port()),
-                "http.target": "/error",
-                "net.peer.ip": "127.0.0.1",
-                "http.status_code": 500,
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_SCHEME: "http",
+                SpanAttributes.HTTP_HOST: "127.0.0.1:"
+                + str(self.get_http_port()),
+                SpanAttributes.HTTP_TARGET: "/error",
+                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_STATUS_CODE: 500,
             },
         )
 
@@ -248,9 +261,9 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             client,
             {
-                "http.url": self.get_url("/error"),
-                "http.method": "GET",
-                "http.status_code": 500,
+                SpanAttributes.HTTP_URL: self.get_url("/error"),
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_STATUS_CODE: 500,
             },
         )
 
@@ -267,12 +280,13 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             server,
             {
-                "http.method": "GET",
-                "http.scheme": "http",
-                "http.host": "127.0.0.1:" + str(self.get_http_port()),
-                "http.target": "/missing-url",
-                "net.peer.ip": "127.0.0.1",
-                "http.status_code": 404,
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_SCHEME: "http",
+                SpanAttributes.HTTP_HOST: "127.0.0.1:"
+                + str(self.get_http_port()),
+                SpanAttributes.HTTP_TARGET: "/missing-url",
+                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_STATUS_CODE: 404,
             },
         )
 
@@ -281,9 +295,9 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             client,
             {
-                "http.url": self.get_url("/missing-url"),
-                "http.method": "GET",
-                "http.status_code": 404,
+                SpanAttributes.HTTP_URL: self.get_url("/missing-url"),
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_STATUS_CODE: 404,
             },
         )
 
@@ -310,12 +324,13 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             server,
             {
-                "http.method": "GET",
-                "http.scheme": "http",
-                "http.host": "127.0.0.1:" + str(self.get_http_port()),
-                "http.target": "/dyna",
-                "net.peer.ip": "127.0.0.1",
-                "http.status_code": 202,
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_SCHEME: "http",
+                SpanAttributes.HTTP_HOST: "127.0.0.1:"
+                + str(self.get_http_port()),
+                SpanAttributes.HTTP_TARGET: "/dyna",
+                SpanAttributes.NET_PEER_IP: "127.0.0.1",
+                SpanAttributes.HTTP_STATUS_CODE: 202,
             },
         )
 
@@ -326,9 +341,9 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             client,
             {
-                "http.url": self.get_url("/dyna"),
-                "http.method": "GET",
-                "http.status_code": 202,
+                SpanAttributes.HTTP_URL: self.get_url("/dyna"),
+                SpanAttributes.HTTP_METHOD: "GET",
+                SpanAttributes.HTTP_STATUS_CODE: 202,
             },
         )
 
@@ -345,9 +360,9 @@ class TestTornadoInstrumentation(TornadoTest):
             self.assert_span_has_attributes(
                 client,
                 {
-                    "http.url": self.get_url(path),
-                    "http.method": "GET",
-                    "http.status_code": 200,
+                    SpanAttributes.HTTP_URL: self.get_url(path),
+                    SpanAttributes.HTTP_METHOD: "GET",
+                    SpanAttributes.HTTP_STATUS_CODE: 200,
                 },
             )
             self.memory_exporter.clear()
@@ -364,6 +379,85 @@ class TestTornadoInstrumentation(TornadoTest):
         self.assert_span_has_attributes(
             server_span, {"uri": "/pong?q=abc&b=123", "query": "q=abc&b=123"}
         )
+        self.memory_exporter.clear()
+
+    def test_response_headers(self):
+        orig = get_global_response_propagator()
+        set_global_response_propagator(TraceResponsePropagator())
+
+        response = self.fetch("/")
+        headers = response.headers
+
+        spans = self.sorted_spans(self.memory_exporter.get_finished_spans())
+        self.assertEqual(len(spans), 3)
+        server_span = spans[1]
+
+        self.assertIn("traceresponse", headers)
+        self.assertEqual(
+            headers["access-control-expose-headers"], "traceresponse",
+        )
+        self.assertEqual(
+            headers["traceresponse"],
+            "00-{0}-{1}-01".format(
+                trace.format_trace_id(server_span.get_span_context().trace_id),
+                trace.format_span_id(server_span.get_span_context().span_id),
+            ),
+        )
+
+        self.memory_exporter.clear()
+        set_global_response_propagator(orig)
+
+
+class TornadoHookTest(TornadoTest):
+    _client_request_hook = None
+    _client_response_hook = None
+    _server_request_hook = None
+
+    def client_request_hook(self, span, handler):
+        if self._client_request_hook is not None:
+            self._client_request_hook(span, handler)
+
+    def client_response_hook(self, span, handler):
+        if self._client_response_hook is not None:
+            self._client_response_hook(span, handler)
+
+    def server_request_hook(self, span, handler):
+        if self._server_request_hook is not None:
+            self._server_request_hook(span, handler)
+
+    def test_hooks(self):
+        def server_request_hook(span, handler):
+            span.update_name("name from server hook")
+            handler.set_header("hello", "world")
+
+        def client_request_hook(span, request):
+            span.update_name("name from client hook")
+
+        def client_response_hook(span, request):
+            span.set_attribute("attr-from-hook", "value")
+
+        self._server_request_hook = server_request_hook
+        self._client_request_hook = client_request_hook
+        self._client_response_hook = client_response_hook
+
+        response = self.fetch("/")
+        self.assertEqual(response.headers.get("hello"), "world")
+
+        spans = self.sorted_spans(self.memory_exporter.get_finished_spans())
+        self.assertEqual(len(spans), 3)
+        server_span = spans[1]
+        self.assertEqual(server_span.kind, SpanKind.SERVER)
+        self.assertEqual(server_span.name, "name from server hook")
+        self.assert_span_has_attributes(server_span, {"uri": "/"})
+        self.memory_exporter.clear()
+
+        client_span = spans[2]
+        self.assertEqual(client_span.kind, SpanKind.CLIENT)
+        self.assertEqual(client_span.name, "name from client hook")
+        self.assert_span_has_attributes(
+            client_span, {"attr-from-hook": "value"}
+        )
+
         self.memory_exporter.clear()